fix: interface cast
I noticed that when implementing multiple interfaces for a type, the struct fields are all messed up

example:
```rust
wrap_window_delegate! {
  struct AppWindowDelegate {
    a: u32,
  }

  impl ViewDelegate {
    fn minimum_size(&self, view: Option<&mut View>) -> cef::Size {
      println!("min size {:?}", self.a);
      cef::Size { width: 0, height: 0 }
    }
  }

  impl PanelDelegate {}

  impl WindowDelegate {
    fn on_window_created(&self, window: Option<&mut Window>) {
      println!(" on window created {:?}", self.a);
      if let Some(window) = window {
        window.show();
      }
    }
  }
}
```
in this case, the WindowDelegate impl works fine, but the ViewDelegate impl does not get the proper `a` value. Looks like that's because the binding is missing a cast when calling a method from a "parent interface".